requirements.txt
# Core dependencies
requests>=2.25.1
beautifulsoup4>=4.9.3
pandas>=1.2.0
matplotlib>=3.3.0
python-dotenv>=0.15.0

# Development dependencies
pylint>=2.7.0
black>=21.5b2
mypy>=0.812
